arby-backend/requirements.txt
Flask>=2.0.2,<3.0
Werkzeug>=2.0.3,<3.0
flask-cors>=3.0.10,<4.0
requests>=2.26.0,<3.0
gunicorn>=20.1.0,<21.0
python-dotenv>=0.19.0,<1.0


